Fix: direct changes and remove interfaces
