--- src/main/java/sonarquberepair/MainApi.java
@@ -1,7 +1,0 @@
-package sonarquberepair;
-
-public interface MainApi {
-	void repair(String pathToFile, String projectKey, int ruleKey, PrettyPrintingStrategy prettyPrintingStrategy) throws Exception;
-
-	void start(String[] args) throws Exception;
-}
