--- src/main/java/sonarquberepair/branch/IRepair.java
@@ -1,7 +1,0 @@
-package sonarquberepair;
-
-/* Different repair behaviours */
-public interface IRepair {
-
-	void repair() throws Exception;
-}
