--- src/test/java/sonarquberepair/branchprocessor/sonarbased/BranchUnclosedResourcesProcessorTest.java
@@ -1,32 +1,0 @@
-package sonarquberepair.processor.sonarbased;
-
-import org.junit.Test;
-import org.sonar.java.checks.verifier.JavaCheckVerifier;
-import org.sonar.java.se.checks.UnclosedResourcesCheck;
-import sonarquberepair.Constants;
-import sonarquberepair.Main;
-import sonarquberepair.TestHelper;
-
-public class BranchUnclosedResourcesProcessorTest {
-
-	@Test
-	public void test() throws Exception {
-
-		String fileName = "ZipFolder.java";
-		String pathToBuggyFile = Constants.PATH_TO_FILE + fileName;
-		String workspace = "sonar-branch-workspace";
-		String pathToRepairedFile = workspace + "/spooned/" + fileName;
-
-		JavaCheckVerifier.verify(pathToBuggyFile, new UnclosedResourcesCheck());
-		Main.main(new String[]{
-			"--versionMode","NEW",
-			"--repairPath",pathToBuggyFile,
-			"--projectKey",Constants.PROJECT_KEY,
-			"--ruleNumbers","2095",
-			"--workspace",workspace});
-		TestHelper.removeComplianceComments(pathToRepairedFile);
-		JavaCheckVerifier.verifyNoIssue(pathToRepairedFile, new UnclosedResourcesCheck());
-
-	}
-
-}
