+++ src/test/java/sonarquberepair/processor/sonarbased/UnclosedResourcesProcessorTest.java
@@ -5,21 +5,27 @@ import org.sonar.java.checks.verifier.JavaCheckVerifier;
 import org.sonar.java.se.checks.UnclosedResourcesCheck;
 import sonarquberepair.Constants;
 import sonarquberepair.Main;
-import sonarquberepair.PrettyPrintingStrategy;
 import sonarquberepair.TestHelper;
 
 public class UnclosedResourcesProcessorTest {
 
 	@Test
 	public void test() throws Exception {
+
 		String fileName = "ZipFolder.java";
 		String pathToBuggyFile = Constants.PATH_TO_FILE + fileName;
-		String pathToRepairedFile = "./spooned/" + fileName;
+		String workspace = "sonar-branch-workspace";
+		String pathToRepairedFile = workspace + "/spooned/" + fileName;
 
 		JavaCheckVerifier.verify(pathToBuggyFile, new UnclosedResourcesCheck());
-		Main.repair(pathToBuggyFile, Constants.PROJECT_KEY, 2095, PrettyPrintingStrategy.NORMAL);
+		Main.main(new String[]{
+			"--repairPath",pathToBuggyFile,
+			"--projectKey",Constants.PROJECT_KEY,
+			"--ruleNumbers","2095",
+			"--workspace",workspace});
 		TestHelper.removeComplianceComments(pathToRepairedFile);
 		JavaCheckVerifier.verifyNoIssue(pathToRepairedFile, new UnclosedResourcesCheck());
+
 	}
 
 }
