+++ src/test/java/sonarquberepair/processor/spoonbased/CompareStringsBoxedTypesWithEqualsProcessorTest.java
@@ -5,7 +5,6 @@ import org.sonar.java.checks.CompareStringsBoxedTypesWithEqualsCheck;
 import org.sonar.java.checks.verifier.JavaCheckVerifier;
 import sonarquberepair.Constants;
 import sonarquberepair.Main;
-import sonarquberepair.PrettyPrintingStrategy;
 import sonarquberepair.TestHelper;
 
 public class CompareStringsBoxedTypesWithEqualsProcessorTest {
@@ -14,10 +13,15 @@ public class CompareStringsBoxedTypesWithEqualsProcessorTest {
 	public void test() throws Exception {
 		String fileName = "CompareStringsBoxedTypesWithEquals.java";
 		String pathToBuggyFile = Constants.PATH_TO_FILE + fileName;
-		String pathToRepairedFile = "./spooned/" + fileName;
+		String workspace = "sonar-branch-workspace";
+		String pathToRepairedFile = workspace + "/spooned/" + fileName;
 
 		JavaCheckVerifier.verify(pathToBuggyFile, new CompareStringsBoxedTypesWithEqualsCheck());
-		Main.repair(pathToBuggyFile, Constants.PROJECT_KEY, 4973, PrettyPrintingStrategy.NORMAL);
+		Main.main(new String[]{
+			"--repairPath",pathToBuggyFile,
+			"--projectKey",Constants.PROJECT_KEY,
+			"--ruleNumbers","4973",
+			"--workspace",workspace});
 		TestHelper.removeComplianceComments(pathToRepairedFile);
 		JavaCheckVerifier.verifyNoIssue(pathToRepairedFile, new CompareStringsBoxedTypesWithEqualsCheck());
 	}
